+++ .gitignore
@@ -3,4 +3,6 @@
 ./__pycache__*
 ./__pycache__/*
 __pycache__/*
+__pycache__/
+
 .env
