+++ app/schemas/user.py
@@ -1,3 +1,5 @@
+# app/schemas/user.py
+
 from pydantic import BaseModel
 from typing import List
 from .room import Room
